gather/pyproject.toml
[project]
name = "gather"
version = "0.1.0"
description = "Support ticket data gathering from Jira, Slack, and other services"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "click>=8.0",
]

[project.scripts]
gather = "gather.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gather"]

[dependency-groups]
dev = [
    "coverage>=7.13.4",
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=7.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
